proxy-server/requirements.txt
python-dotenv==1.1.0
fastapi==0.115.12
httpx==0.28.1
pydantic==2.11.3
pydantic-ai==0.1.6
